Reorganize ServiceTest
