+++ tests/Unit/Config/ServiceTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Payavel\Orchestration\Tests\Unit;
+namespace Payavel\Orchestration\Tests\Unit\Config;
 
 use Payavel\Orchestration\Tests\Traits\CreatesConfigServiceables;
 use Payavel\Orchestration\Tests\Traits\SetsConfigDriver;
+use Payavel\Orchestration\Tests\Unit\TestService;
 
-class ConfigServiceTest extends TestService
+class ServiceTest extends TestService
 {
     use CreatesConfigServiceables;
     use SetsConfigDriver;
