+++ tests/Unit/Database/ServiceTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Payavel\Orchestration\Tests\Unit;
+namespace Payavel\Orchestration\Tests\Unit\Database;
 
 use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
 use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
+use Payavel\Orchestration\Tests\Unit\TestService;
 
-class DatabaseServiceTest extends TestService
+class ServiceTest extends TestService
 {
     use CreatesDatabaseServiceables;
     use SetsDatabaseDriver;
